chore: Updated type hints
