+++ fastembed/common/types.py
@@ -1,7 +1,8 @@
 from pathlib import Path
 import sys
 from PIL import Image
-from typing import Any, Iterable, Union
+from dataclasses import dataclass
+from typing import Any, Iterable, Union, Optional, TypeVar
 
 import numpy as np
 from numpy.typing import NDArray
@@ -11,6 +12,8 @@ if sys.version_info >= (3, 10):
 else:
     from typing_extensions import TypeAlias
 
+# Holds type of the embedding result
+T = TypeVar("T")
 
 PathInput: TypeAlias = Union[str, Path]
 PilInput: TypeAlias = Union[Image.Image, Iterable[Image.Image]]
@@ -25,3 +28,10 @@ NumpyArray = Union[
     NDArray[np.int64],
     NDArray[np.int32],
 ]
+
+
+@dataclass
+class OnnxOutputContext:
+    model_output: NumpyArray
+    attention_mask: Optional[NumpyArray] = None
+    input_ids: Optional[NumpyArray] = None
