+++ fastembed/text/clip_embedding.py
@@ -1,7 +1,6 @@
 from typing import Any, Iterable, Type
 
-from fastembed.common.types import NumpyArray
-from fastembed.common.onnx_model import OnnxOutputContext
+from fastembed.common.types import NumpyArray, OnnxOutputContext
 from fastembed.text.onnx_embedding import OnnxTextEmbedding, OnnxTextEmbeddingWorker
 
 supported_clip_models = [
