+++ fastembed/text/pooled_embedding.py
@@ -2,8 +2,7 @@ from typing import Any, Iterable, Type
 
 import numpy as np
 
-from fastembed.common.types import NumpyArray
-from fastembed.common.onnx_model import OnnxOutputContext
+from fastembed.common.types import NumpyArray, OnnxOutputContext
 from fastembed.text.onnx_embedding import OnnxTextEmbedding, OnnxTextEmbeddingWorker
 
 supported_pooled_models = [
