+++ fastembed/text/text_embedding.py
@@ -1,8 +1,7 @@
 import warnings
 from typing import Any, Iterable, Optional, Sequence, Type, Union
 
-from fastembed.common.types import NumpyArray
-from fastembed.common import OnnxProvider
+from fastembed.common.types import NumpyArray, OnnxProvider
 from fastembed.text.clip_embedding import CLIPOnnxEmbedding
 from fastembed.text.pooled_normalized_embedding import PooledNormalizedEmbedding
 from fastembed.text.pooled_embedding import PooledEmbedding
